# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mock AI API",
    "version": "0.1.0",
    "description": "用于 AI API 开发、集成测试和性能测试的多协议 Mock 服务。所有 API Key 仅用于模拟客户端请求，服务不会校验其内容。流式接口可通过请求体中的 `stream` 字段启用，并支持 `mock_ttft_ms`、`mock_tps`、`mock_latency_ms`、`mock_input_tokens` 和 `mock_output_tokens` 测试参数。",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [
    { "url": "/", "description": "当前 Mock AI API 服务" }
  ],
  "security": [],
...
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": ["通用"],
        "summary": "获取模型列表",
        "description": "默认返回 OpenAI 格式；传入 `anthropic-version` 返回 Anthropic 格式；传入 `x-goog-api-key` 或 `key` 返回 Gemini 格式。",
        "operationId": "listModels",
        "parameters": [
          { "$ref": "#/components/parameters/AnthropicVersion" },
          { "$ref": "#/components/parameters/GeminiKey" }
        ],
        "security": [
          { "BearerAuth": [] },
          { "AnthropicApiKey": [] },
          { "GeminiApiKeyHeader": [] },
          {}
        ],
        "responses": { "200": { "$ref": "#/components/responses/JSONResponse" }, "400": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/v1/models/{model}": {
      "get": {
        "tags": ["通用"],
        "summary": "获取模型详情",
        "description": "响应协议识别规则与模型列表接口相同。",
        "operationId": "getModel",
        "parameters": [
          { "$ref": "#/components/parameters/Model" },
          { "$ref": "#/components/parameters/AnthropicVersion" },
          { "$ref": "#/components/parameters/GeminiKey" }
        ],
        "security": [
          { "BearerAuth": [] },
          { "AnthropicApiKey": [] },
          { "GeminiApiKeyHeader": [] },
          {}
        ],
        "responses": { "200": { "$ref": "#/components/responses/JSONResponse" }, "400": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": ["OpenAI"],
        "summary": "创建聊天补全",
        "operationId": "createChatCompletion",
        "security": [{ "BearerAuth": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAIChat" },
        "responses": {
          "200": {
            "description": "非流式 JSON 或 SSE 流",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ArbitraryObject" } },
              "text/event-stream": { "schema": { "type": "string" }, "example": "data: {\"object\":\"chat.completion.chunk\"}\n\ndata: [DONE]\n\n" }
            }
          },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/v1/completions": {
      "post": {
        "tags": ["OpenAI"],
        "summary": "创建文本补全",
        "operationId": "createCompletion",
        "security": [{ "BearerAuth": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAICompletion" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONOrStreamResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/v1/embeddings": {
      "post": {
        "tags": ["OpenAI"],
        "summary": "创建向量",
        "operationId": "createEmbedding",
        "security": [{ "BearerAuth": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAIEmbedding" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/v1/responses": {
      "post": {
        "tags": ["OpenAI"],
        "summary": "创建响应",
        "operationId": "createResponse",
        "security": [{ "BearerAuth": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAIResponse" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONOrStreamResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/v1/messages": {
      "post": {
        "tags": ["Anthropic"],
        "summary": "创建消息",
        "operationId": "createAnthropicMessage",
        "parameters": [{ "$ref": "#/components/parameters/AnthropicVersionRequired" }],
        "security": [{ "AnthropicApiKey": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/AnthropicMessage" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONOrStreamResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/v1/messages/count_tokens": {
      "post": {
        "tags": ["Anthropic"],
        "summary": "计算输入 Token",
        "operationId": "countAnthropicTokens",
        "parameters": [{ "$ref": "#/components/parameters/AnthropicVersionRequired" }],
        "security": [{ "AnthropicApiKey": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/AnthropicMessage" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/openai/deployments/{deployment}/chat/completions": {
      "post": {
        "tags": ["Azure OpenAI"],
        "summary": "创建聊天补全",
        "operationId": "createAzureChatCompletion",
        "parameters": [
          { "$ref": "#/components/parameters/Deployment" },
          { "$ref": "#/components/parameters/APIVersion" }
        ],
        "security": [{ "AzureApiKey": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAIChat" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONOrStreamResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/openai/deployments/{deployment}/completions": {
      "post": {
        "tags": ["Azure OpenAI"],
        "summary": "创建文本补全",
        "operationId": "createAzureCompletion",
        "parameters": [
          { "$ref": "#/components/parameters/Deployment" },
          { "$ref": "#/components/parameters/APIVersion" }
        ],
        "security": [{ "AzureApiKey": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAICompletion" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONOrStreamResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/openai/deployments/{deployment}/embeddings": {
      "post": {
        "tags": ["Azure OpenAI"],
        "summary": "创建向量",
        "operationId": "createAzureEmbedding",
        "parameters": [
          { "$ref": "#/components/parameters/Deployment" },
          { "$ref": "#/components/parameters/APIVersion" }
        ],
        "security": [{ "AzureApiKey": [] }, {}],
        "requestBody": { "$ref": "#/components/requestBodies/OpenAIEmbedding" },
        "responses": {
          "200": { "$ref": "#/components/responses/JSONResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/api/embeddings": {
      "post": {
        "tags": ["Ollama"], "summary": "创建向量（旧版）", "operationId": "createOllamaLegacyEmbedding",
        "requestBody": { "$ref": "#/components/requestBodies/OllamaLegacyEmbedding" },
        "responses": { "200": { "$ref": "#/components/responses/JSONResponse" }, "400": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "API Key", "description": "OpenAI API Key；Mock 服务不校验内容。" },
      "AzureApiKey": { "type": "apiKey", "in": "header", "name": "api-key", "description": "Azure OpenAI API Key；Mock 服务不校验内容。" },
      "AnthropicApiKey": { "type": "apiKey", "in": "header", "name": "x-api-key", "description": "Anthropic API Key；Mock 服务不校验内容。" },
      "GeminiApiKeyHeader": { "type": "apiKey", "in": "header", "name": "x-goog-api-key", "description": "Gemini API Key 请求头；Mock 服务不校验内容。" },
      "GeminiApiKeyQuery": { "type": "apiKey", "in": "query", "name": "key", "description": "Gemini API Key 查询参数；Mock 服务不校验内容。" }
    },
    "parameters": {
      "Model": { "name": "model", "in": "path", "required": true, "description": "模型名称", "schema": { "type": "string", "default": "mock-gpt" } },
      "Deployment": { "name": "deployment", "in": "path", "required": true, "description": "Azure deployment 名称", "schema": { "type": "string", "default": "mock-gpt" } },
      "APIVersion": { "name": "api-version", "in": "query", "required": true, "description": "Azure OpenAI API 版本", "schema": { "type": "string", "default": "2024-10-21" } },
      "AnthropicVersion": { "name": "anthropic-version", "in": "header", "required": false, "description": "设置后共享模型接口返回 Anthropic 格式", "schema": { "type": "string", "example": "2023-06-01" } },
      "AnthropicVersionRequired": { "name": "anthropic-version", "in": "header", "required": true, "description": "Anthropic API 版本", "schema": { "type": "string", "default": "2023-06-01" } },
      "GeminiKey": { "name": "key", "in": "query", "required": false, "description": "Gemini API Key；可填写任意测试值", "schema": { "type": "string", "example": "test" } },
...
    },
    "requestBodies": {
      "OpenAIChat": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenAIChatRequest" } } }
      },
      "OpenAICompletion": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenAICompletionRequest" } } }
      },
      "OpenAIResponse": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenAIResponseRequest" } } }
      },
      "OpenAIEmbedding": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenAIEmbeddingRequest" } } }
      },
      "AnthropicMessage": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnthropicMessageRequest" } } }
      },
...
      "OllamaLegacyEmbedding": {
        "required": true,
        "content": { "application/json": { "schema": { "type": "object", "properties": { "model": { "type": "string", "default": "mock-gpt" }, "prompt": { "type": "string", "default": "你好" }, "mock_input_tokens": { "$ref": "#/components/schemas/MockInputTokens" } }, "required": ["prompt"] } } }
      }
    },
    "responses": {
      "JSONResponse": { "description": "JSON 响应", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArbitraryObject" } } } },
...
      "JSONOrStreamResponse": { "description": "非流式 JSON 或 SSE 流", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArbitraryObject" } }, "text/event-stream": { "schema": { "type": "string" } } } },
...
      "ErrorResponse": { "description": "请求参数错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArbitraryObject" } } } }
    },
    "schemas": {
      "ArbitraryObject": { "type": "object", "additionalProperties": true },
      "Message": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "enum": ["system", "user", "assistant"], "default": "user" },
          "content": { "oneOf": [{ "type": "string" }, { "type": "array", "items": {} }], "default": "你好" }
        },
        "required": ["role", "content"]
      },
      "Stream": { "type": "boolean", "default": false, "description": "是否返回流式响应" },
      "MockTTFT": { "type": "integer", "minimum": 0, "maximum": 3600000, "default": 100, "description": "首个文本 Token 等待毫秒数" },
      "MockTPS": { "type": "number", "minimum": 0, "maximum": 1000000, "default": 20, "description": "流式输出 Token 每秒；0 表示不等待" },
      "MockLatency": { "type": "integer", "minimum": 0, "maximum": 3600000, "default": 50, "description": "非流式响应延迟毫秒数" },
      "MockInputTokens": { "type": "integer", "minimum": 0, "maximum": 1000000000, "default": 8, "description": "Usage 中的输入 Token 数" },
      "MockOutputTokens": { "type": "integer", "minimum": 1, "maximum": 1000000, "default": 8, "description": "实际生成及 Usage 中的输出 Token 数" },
      "OpenAIChatRequest": {
        "type": "object",
        "properties": {
          "model": { "type": "string", "default": "mock-gpt" },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/Message" }, "default": [{ "role": "user", "content": "你好" }] },
          "stream": { "$ref": "#/components/schemas/Stream" },
          "max_tokens": { "type": "integer", "minimum": 1, "default": 32 },
          "max_completion_tokens": { "type": "integer", "minimum": 1 },
          "stream_options": { "type": "object", "additionalProperties": true },
          "mock_ttft_ms": { "$ref": "#/components/schemas/MockTTFT" },
          "mock_tps": { "$ref": "#/components/schemas/MockTPS" },
          "mock_latency_ms": { "$ref": "#/components/schemas/MockLatency" },
          "mock_input_tokens": { "$ref": "#/components/schemas/MockInputTokens" },
          "mock_output_tokens": { "$ref": "#/components/schemas/MockOutputTokens" }
        },
        "required": ["messages"]
      },
      "OpenAICompletionRequest": {
        "type": "object",
        "properties": {
          "model": { "type": "string", "default": "mock-gpt" },
          "prompt": { "oneOf": [{ "type": "string" }, { "type": "array", "items": { "type": "string" } }], "default": "你好" },
          "stream": { "$ref": "#/components/schemas/Stream" },
          "max_tokens": { "type": "integer", "minimum": 1, "default": 32 },
          "mock_ttft_ms": { "$ref": "#/components/schemas/MockTTFT" },
          "mock_tps": { "$ref": "#/components/schemas/MockTPS" },
          "mock_latency_ms": { "$ref": "#/components/schemas/MockLatency" },
          "mock_input_tokens": { "$ref": "#/components/schemas/MockInputTokens" },
          "mock_output_tokens": { "$ref": "#/components/schemas/MockOutputTokens" }
        },
        "required": ["prompt"]
      },
      "OpenAIResponseRequest": {
        "type": "object",
        "properties": {
          "model": { "type": "string", "default": "mock-gpt" },
          "input": { "oneOf": [{ "type": "string" }, { "type": "array", "items": {} }], "default": "你好" },
          "stream": { "$ref": "#/components/schemas/Stream" },
          "max_output_tokens": { "type": "integer", "minimum": 1, "default": 32 },
          "mock_ttft_ms": { "$ref": "#/components/schemas/MockTTFT" },
          "mock_tps": { "$ref": "#/components/schemas/MockTPS" },
          "mock_latency_ms": { "$ref": "#/components/schemas/MockLatency" },
          "mock_input_tokens": { "$ref": "#/components/schemas/MockInputTokens" },
          "mock_output_tokens": { "$ref": "#/components/schemas/MockOutputTokens" }
        },
        "required": ["input"]
      },
      "OpenAIEmbeddingRequest": {
        "type": "object",
        "properties": {
          "model": { "type": "string", "default": "mock-gpt" },
          "input": { "oneOf": [{ "type": "string" }, { "type": "array", "items": {} }], "default": "你好" },
          "mock_latency_ms": { "$ref": "#/components/schemas/MockLatency" },
          "mock_input_tokens": { "$ref": "#/components/schemas/MockInputTokens" }
        },
        "required": ["input"]
      },
      "AnthropicMessageRequest": {
        "type": "object",
        "properties": {
          "model": { "type": "string", "default": "mock-gpt" },
          "max_tokens": { "type": "integer", "minimum": 1, "default": 32 },
          "system": { "oneOf": [{ "type": "string" }, { "type": "array", "items": {} }] },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/Message" }, "default": [{ "role": "user", "content": "你好" }] },
          "stream": { "$ref": "#/components/schemas/Stream" },
          "mock_ttft_ms": { "$ref": "#/components/schemas/MockTTFT" },
          "mock_tps": { "$ref": "#/components/schemas/MockTPS" },
          "mock_latency_ms": { "$ref": "#/components/schemas/MockLatency" },
          "mock_input_tokens": { "$ref": "#/components/schemas/MockInputTokens" },
          "mock_output_tokens": { "$ref": "#/components/schemas/MockOutputTokens" }
        },
        "required": ["messages"]
      },
...
    }
  }
}